+++ main.py
@@ -1,0 +1,49 @@
+# Author: <Chris Ward>
+# Creation Date: <05/01/2021>
+# Below is a simple program with 10 issues (some are syntax errors and some are logic errors.  You need to identify the issues and correct them.
+
+import random
+import time
+
+def displayIntro():
+  print("\nYou are in a land full of dragons. In front of you,you see two caves. In one cave, the dragon is friendly and will share his treasure with you. The other dragon is greedy and hungry, and will eat you on sight.")
+#displayIntro()
+#cleaned up intro
+
+#TabError: inconsistent use of tabs and spaces in indentation cave != '2':
+def chooseCave():
+  cave = ''
+  while cave != '1' and cave != '2':
+    print('\nWhich cave will you go into? (1 or 2)')
+    cave = input()
+  return cave
+
+def checkCave(chosenCave):
+  print('\nYou approach the cave...')
+  #sleep for 2 seconds
+  time.sleep(2)
+  print('\nIt is dark and spooky...')
+  #sleep for 2 seconds
+  time.sleep(3)
+  print('\nA large dragon jumps out in front of you! He opens his jaws and...')
+  print()
+  #sleep for 2 seconds
+  time.sleep(2)
+  friendlyCave = random.randint(1, 2)
+
+  if chosenCave == str(friendlyCave):
+    print('\n\nGives you his treasure!')
+  else:
+    print ('\nGobbles you down in one bite!')
+
+playAgain = 'yes'
+while playAgain == 'yes' or playAgain == 'y':
+  displayIntro()
+  caveNumber = chooseCave()
+  checkCave(caveNumber)
+    
+  print("\nDo you want to play again? (Enter 'y' or 'yes' or any key to exit)")
+  playAgain = input()
+  #if playAgain == "no" or playAgain == "n":
+if playAgain != "yes" or playAgain != "y":
+  print("Thanks for playing!")
